requirements.txt
# Framework Principal e Extensões
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
Flask-Mail
Flask-APScheduler

# Biblioteca da IA (versão mínima para garantir a sintaxe correta)
google-generativeai>=0.3

# ORM (versão mínima para garantir a sintaxe correta)
SQLAlchemy>=2.0

# Utilitários
python-dotenv

email_validator

